+++ R/AcqOptimizerLbfgsb.R
@@ -22,6 +22,11 @@
 #' \item{`max_restarts`}{`integer(1)`\cr
 #'   Maximum number of restarts.
 #'   Default is `5 * D` (Default).}
+#' \item{`skip_already_evaluated`}{`logical(1)`\cr
+#'   Should the proposed candidate be rejected if it was already evaluated on the actual [bbotk::OptimInstance]?
+#'   If `TRUE` and the candidate was already evaluated, an error is raised so that the `loop_function` can
+#'   propose a randomly sampled point instead.
+#'   Default is `TRUE`.}
 #' }
 #'
 #' @note
@@ -86,6 +91,7 @@ AcqOptimizerLbfgsb = R6Class(
         ftol_abs = p_dbl(default = 0, lower = 0, upper = Inf, special_vals = list(-1)),
         restart_strategy = p_fct(levels = c("none", "random"), init = "none"),
         max_restarts = p_int(lower = 0L),
+        skip_already_evaluated = p_lgl(init = TRUE),
         catch_errors = p_lgl(init = TRUE)
       )
       private$.param_set = param_set
@@ -105,10 +111,12 @@ AcqOptimizerLbfgsb = R6Class(
       max_restarts = pv$max_restarts
       maxeval = pv$maxeval
       catch_errors = pv$catch_errors
+      skip_already_evaluated = pv$skip_already_evaluated
       pv$max_restarts = NULL
       pv$restart_strategy = NULL
       pv$maxeval = NULL
       pv$catch_errors = NULL
+      pv$skip_already_evaluated = NULL
 
       if (restart_strategy == "none") {
         max_restarts = 0L
@@ -207,10 +215,14 @@ AcqOptimizerLbfgsb = R6Class(
       if (is.null(x)) {
         error_acq_optimizer("Acquisition function optimization did not yield a valid solution.")
       }
-      as.data.table(as.list(set_names(
+      xdt = as.data.table(as.list(set_names(
         c(x, y * direction),
         c(self$acq_function$domain$ids(), self$acq_function$codomain$ids())
       )))
+      if (skip_already_evaluated) {
+        assert_not_already_evaluated(xdt, self$acq_function$archive)
+      }
+      xdt
     },
 
     #' @description